Fix to prevent image edit and delete by user other than image owner

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,12 +92,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");  // Get user from logged in session
 
+        Image image = imageService.getImage(imageId);
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
+
+        if(image.getUser().getId() != user.getId()) { // Validate image owner against logged-in user
+            Boolean editError = true;
+            model.addAttribute("editError", editError);  // Add "editError" attribute
+            return "images/image";                        // Return to the image.html
+        }
+
         return "images/edit";
     }
 
@@ -140,7 +148,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        User user = (User) session.getAttribute("loggeduser");   // Get user from logged in session
+        Image image = imageService.getImage(imageId);
+
+        if(image.getUser().getId() != user.getId()) {                // Validate image owner against logged-in user
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+
+            Boolean deleteError = true;
+            model.addAttribute("deleteError", deleteError);   // Add "deleteError" attribute to the model
+            return "images/image";                                  // Return to the image.html
+        }
+
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
